Refactoring: Move fixtures in test_install.py into conftest.py
To split linux distribution specific test files from test_install.py later.
